refactor: Add request and response in dto package (#6)
Added request and response subpackages in dto package.
Included DTOs for request in request package and included DTOs for response in response package.
