+++ backend/Airbnb/src/main/java/com/codesquad/team10/airbnb/dto/request/LogInDto.java
@@ -1,4 +1,4 @@
-package com.codesquad.team10.airbnb.dto;
+package com.codesquad.team10.airbnb.dto.request;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
