+++ backend/Airbnb/src/main/java/com/codesquad/team10/airbnb/dto/response/SearchResultDto.java
@@ -1,4 +1,4 @@
-package com.codesquad.team10.airbnb.dto;
+package com.codesquad.team10.airbnb.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
